dyliveapp/requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-simpleui>=2023.3.1
django-filter>=23.5
mysqlclient>=2.2.0
django-cors-headers>=4.3.1
Pillow>=10.0.0
djangorestframework-simplejwt>=5.3.0
argon2-cffi>=21.3.0

# Testing dependencies
pytest>=8.0.0
pytest-django>=4.7.0
pytest-xdist>=3.5.0
pytest-cov>=4.1.0
coverage>=7.4.0